+++ crates/ruff_linter/src/rules/pycodestyle/rules/not_tests.rs
@@ -1,12 +1,12 @@
 use crate::fix::edits::pad;
 use ruff_diagnostics::{AlwaysFixableViolation, Diagnostic, Edit, Fix};
 use ruff_macros::{derive_message_formats, violation};
+use ruff_python_ast::helpers::generate_comparison;
 use ruff_python_ast::{self as ast, CmpOp, Expr};
 use ruff_text_size::Ranged;
 
 use crate::checkers::ast::Checker;
 use crate::registry::{AsRule, Rule};
-use crate::rules::pycodestyle::helpers::generate_comparison;
 
 /// ## What it does
 /// Checks for negative comparison using `not {foo} in {bar}`.
